feat(sheet): group document optimized
