+++ src/components/editor/operator/sheet/SheetGroup.tsx
@@ -1,26 +1,17 @@
-import {
-  Button,
-  Card,
-  Collapse,
-  Divider,
-  Icon,
-  NonIdealState,
-} from '@blueprintjs/core'
+import { Divider, NonIdealState } from '@blueprintjs/core'
 
 import { useAtom } from 'jotai'
-import { useMemo, useState } from 'react'
+import { useMemo } from 'react'
 import { UseFieldArrayRemove, UseFormSetError } from 'react-hook-form'
 
-import { CardDeleteOption } from 'components/editor/CardOptions'
 import { CopilotDocV1 } from 'models/copilot.schema'
 import { OPERATORS, PROFESSIONS } from 'models/generated/operators'
 import { favGroupAtom } from 'store/useFavGroups'
 
 import { EditorPerformerGroupProps } from '../EditorPerformerGroup'
 import { SheetContainerSkeleton } from './SheetContainerSkeleton'
-import { OperatorItem } from './SheetOperatorItem'
+import { EventType, Group, GroupItem } from './sheetGroup/SheetGroupItem'
 
-type Group = CopilotDocV1.Group
 type Operators = CopilotDocV1.Operator[]
 
 export interface SheetGroupProps {
@@ -29,8 +20,6 @@ export interface SheetGroupProps {
   existedOperators: Operators
   removeGroup: UseFieldArrayRemove
 }
-
-type EventType = 'add' | 'remove' | 'pin' | 'opers'
 
 const SheetGroup = ({
   submitGroup,
@@ -47,11 +36,14 @@ const SheetGroup = ({
             OPERATORS.find((opInfoItem) => opInfoItem.name === opItem.name)
               ?.subProf === subProItem.id,
         )
-        if (operators.length)
-          result.push({
-            name: proItem.name + '-' + subProItem.name,
-            opers: operators,
-          })
+        if (operators.length) {
+          const groupName = proItem.name + '-' + subProItem.name
+          if (!existedGroups.find(({ name }) => name === groupName))
+            result.push({
+              name: groupName,
+              opers: operators,
+            })
+        }
       })
     })
     const otherOperators = existedOperators.filter(
@@ -157,100 +149,5 @@ export const SheetGroupContainer = (sheetGroupProps: SheetGroupProps) => (
     <SheetGroup {...sheetGroupProps} />
   </SheetContainerSkeleton>
 )
-
-interface GroupItemProps {
-  groupInfo: Group
-  editable: boolean
-  exist: boolean
-  pinned: boolean
-  eventHandleProxy: (
-    type: EventType,
-    value: Group,
-    setError?: UseFormSetError<CopilotDocV1.Operator>,
-  ) => void
-}
-
-const GroupItem = ({
-  groupInfo,
-  editable,
-  exist,
-  pinned,
-  eventHandleProxy,
-}: GroupItemProps) => {
-  const [showOperators, setShowOperators] = useState(editable)
-  const clickHandle = () => {
-    if (exist) {
-      if (editable) eventHandleProxy('remove', groupInfo)
-    } else eventHandleProxy('add', groupInfo)
-  }
-
-  const Operators = useMemo(
-    () => (
-      <div className="flex flex-wrap w-full pt-1">
-        {groupInfo.opers?.map((item) => (
-          <div className="w-1/4 p-1">
-            <OperatorItem
-              id={
-                OPERATORS.find((opInfoitem) => opInfoitem.name === item.name)
-                  ?.id || ''
-              }
-              name={item.name}
-              selected={false}
-              interactive={false}
-            />
-          </div>
-        ))}
-        {editable && (
-          <div className="w-1/4 p-1">
-            <Card className="h-full flex items-center justify-center">
-              <Icon icon="plus" size={35} />
-            </Card>
-          </div>
-        )}
-      </div>
-    ),
-    [groupInfo.opers],
-  )
-
-  return (
-    <Card interactive className="mt-1">
-      <div className="flex items-center">
-        <Icon icon="people" />
-        <p className="ml-1">{groupInfo.name}</p>
-        <div className="ml-auto flex items-center">
-          <Button
-            minimal
-            icon={showOperators ? 'remove' : 'add'}
-            title={`${showOperators ? '折叠' : '展开'}所包含干员`}
-            className="cursor-pointer"
-            onClick={() => setShowOperators(!showOperators)}
-          />
-          {editable ? (
-            <CardDeleteOption
-              className="cursor-pointer"
-              onClick={clickHandle}
-            />
-          ) : (
-            <Button
-              minimal
-              icon={exist ? 'tick' : 'arrow-left'}
-              title={exist ? '已选择' : '使用该推荐分组'}
-              onClick={clickHandle}
-            />
-          )}
-          {editable && (
-            <Button
-              minimal
-              icon={pinned ? 'star' : 'star-empty'}
-              title={pinned ? `从收藏移除` : `添加至收藏`}
-              onClick={() => eventHandleProxy('pin', groupInfo)}
-            />
-          )}
-        </div>
-      </div>
-      <Collapse isOpen={showOperators}>{Operators}</Collapse>
-    </Card>
-  )
-}
 
 const GroupNoData = <NonIdealState title="暂无干员组" />
