+++ src/store/useFavGroups.ts
@@ -1,9 +1,18 @@
+import { atom } from 'jotai'
+
 import { CopilotDocV1 } from '../models/copilot.schema'
 import { atomWithLocalStorage } from './storage'
 
 type FavGroupState = CopilotDocV1.Group[]
 
-export const favGroupAtom = atomWithLocalStorage<FavGroupState, FavGroupState>(
+const favGroupCoreAtom = atomWithLocalStorage<FavGroupState, FavGroupState>(
   'maa-copilot-fav-groups',
   [],
+)
+
+export const favGroupAtom = atom(
+  (get) => get(favGroupCoreAtom),
+  (_get, set, value: FavGroupState) => {
+    set(favGroupCoreAtom, value)
+  },
 )
